+++ packages/agent-browser-extension/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @agent-native/agent-browser-extension
 
+## 0.1.199
+
+### Patch Changes
+
+- Updated dependencies [b1c420b]
+- Updated dependencies [8690e40]
+- Updated dependencies [e542242]
+  - @agent-native/core@0.164.9
+
+## 0.1.198
+
+### Patch Changes
+
+- Updated dependencies [939f6d2]
+  - @agent-native/core@0.164.8
+
 ## 0.1.197
 
 ### Patch Changes