+++ packages/mobile-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @agent-native/mobile-app
 
+## 0.1.48
+
+### Patch Changes
+
+- Updated dependencies [b1c420b]
+- Updated dependencies [8690e40]
+- Updated dependencies [e542242]
+  - @agent-native/core@0.164.9
+
+## 0.1.47
+
+### Patch Changes
+
+- Updated dependencies [939f6d2]
+  - @agent-native/core@0.164.8
+
 ## 0.1.46
 
 ### Patch Changes